refactor: converted course schedule to ts
